caduceus: dissolve caduceus type's cauduces handlers into http.go
